backend/src/ocr/requirements.txt
opencv-python>=4.8.0
surya-ocr>=0.17.0
torch>=2.0.0
paddleocr>=2.7.0
paddlepaddle-gpu>=2.5.0
numpy>=1.24.0
pillow>=10.0.0
